==> ADHD_script/S2_normativemodeling/S5_plotNM_eachGroup_ABCD.qmd <==
---
title: "S5_plotNM_eachGroup"
author:
  - name: Xiaoyu Xu
    orcid: 0000-0002-7349-3366
    email: xuxiaoyu@cibr.ac.cn
    url: https://scholar.google.com/citations?user=G0jeI1AAAAAJl
    affiliation: 
      - name: Chinese Institute for Brain Research, Beijing (CIBR)
        group: CUIZAIXU Lab
        url: https://cuilab.cibr.ac.cn/
date: "2025-09-02"
format: 
  html:
    code-fold: true
    code-tools: true
    theme: Materia
    toc: true
    toc_depth: 3
    toc-expand: 1
    toc_float: true
    toc-location: left
    number-sections: true
    number-depth: 3
    embed-resources: true
---

This script visualizes developmental trajectories derived from the normative models and their critical features.
Besides, this script visualizes the average deviations of both groups.

```{r}
#| label: load-packages-data
#| include: false
library(ggplot2)
library(tidyverse)
library(mgcv)
library(readxl)
library(openxlsx)
library(parallel)
library(gamlss)
library(scales)
library(corrplot)
library(effsize)
rm(list=ls())
# set resolution
Yeoresolution <- 17
if (Yeoresolution == 7){
  Yeoresolution.delLM = 6
}else if (Yeoresolution == 17){
  Yeoresolution.delLM = 15
}
element_num <- Yeoresolution.delLM*(Yeoresolution.delLM+1)/2

# input directory
wd <- getwd()
homepath <- str_split_i(wd, "Normative_model", 1)
demopath <- paste0(homepath, '/Normative_model/demography')
interfileFolder <- paste0(homepath, '/Normative_model/interfileFolder_ABCD')
functionFolder <- paste0(homepath, "/Normative_model/functions")
resultFolder <- paste0(homepath, "/Normative_model/results_ABCD")
functionFolder.SCDev <- paste0(homepath, "/SC_development/Rcode_SCdevelopment/gamfunction")
FigureFolder <- paste0(homepath, '/Normative_model/Figures_ABCD/Yeo', Yeoresolution,'/CV75')

# load data
SCdata <- readRDS(paste0(interfileFolder, '/SCdata_Yeo',Yeoresolution, '_CV75_sumSCinvnode.sum.msmtcsd.merge.rds'))
SCdata$sex <- as.factor(SCdata$demo_sex_v2)
SCdata <- SCdata %>% mutate(WB_SCmean = rowMeans(dplyr::select(.,starts_with("SC."))))

SCdata.TD.trainset <- readRDS(paste0(interfileFolder, "/SCdata.TD.trainset_SCYeo", element_num, ".rds"))
SCdata.TD.testset <- readRDS(paste0(interfileFolder, "/SCdata.TD.testset_SCYeo", element_num, ".rds"))
SCdata.ADHD <- SCdata %>% filter(if_TD == 0)

mod.TD.sum <- readRDS(paste0(interfileFolder, "/GAMLSS_Yeo", Yeoresolution,".TDtraining.sum.rds"))
mod.training.WB <- readRDS(paste0(interfileFolder, "/GAMLSS_Yeo", Yeoresolution,"_WBSCmean.TDtraining.sum.rds"))
source(paste0(functionFolder, "/Construct_gamlss_set.R"))
source(paste0(functionFolder, "/plotmatrix.R"))
source(paste0(functionFolder.SCDev, '/SCrankcorr.R'))

deviationdf <- readRDS(paste0(interfileFolder, '/SCdata_Yeo',Yeoresolution, '_CV75_sumSCinvnode.deviations_TDtest_ADHD.rds'))

```

## Step1 Construct gamlss for ADHD

This step aim to fit the developmental curves for the ADHD group.

```{r}
## Fit normative models
SCdata.ADHD <- as.data.frame(SCdata.ADHD)
SCdata.ADHD[,c("sex", "siteID")] <- lapply(SCdata.ADHD[,c("sex", "siteID")], as.factor)

n0 = nrow(SCdata.ADHD)
SCdata.ADHD <- SCdata.ADHD %>% filter(WB_SCmean>mean(WB_SCmean)-3*sd(WB_SCmean), WB_SCmean<mean(WB_SCmean)+3*sd(WB_SCmean), siteID %in% unique(SCdata.TD.trainset$siteID))
n1 = nrow(SCdata.ADHD)
#boxplot(SCdata.ADHD$WB_SCmean)
print(paste(n1, "obs included,", (n0-n1), "obs were removed due to extreme deviation of WB_SCmean & not in the site list."))

dataname <- "SCdata.ADHD"
smoothterm <- "age"
covariates <- "sex+meanFD"
randomvar <- "siteID"
mu.df <- sigma.df <- degree <- 2
distribution.fam <- "GG"
IDvar <- "scanID"
quantile.vec <- c(0.025, 0.5, 0.975)
stratify <- c("sex", "siteID")
if (! file.exists(paste0(interfileFolder, "/GAMLSS_Yeo", Yeoresolution,".ADHD.sum.rds"))){
  mod.ADHD.sum <- mclapply(1:element_num, function(i){
    dependentvar <- paste0("SC.", i)
    invisible(sumlist <- construct_gamlss(dataname, dependentvar, smoothterm, covariates,randomvar, 
                                mu.df, sigma.df, degree, distribution.fam,IDvar, quantile.vec, stratify))
    
    return(sumlist)
  }, mc.cores = 40)
  saveRDS(mod.ADHD.sum, paste0(interfileFolder, "/GAMLSS_Yeo", Yeoresolution,".ADHD.sum.rds"))
}else{
  mod.ADHD.sum <- readRDS(paste0(interfileFolder, "/GAMLSS_Yeo", Yeoresolution,".ADHD.sum.rds"))
}

performanceADHD.model <- do.call(rbind, lapply(1:element_num, function(x) mod.ADHD.sum[[x]]$performance.tb))
print(paste(sum(performanceADHD.model$converged)*100/element_num, "% models converged."))
# 100 % models converged.


```

## Step2 Compute the 1st derivatives for the P50 trajectories

Use backward finite differences.

```{r}

# define 1st derivative
comuptederivative <- function(smoothvar, mod.tmp, n_points, n_sites){
  # estimate first derivatives (finite difference)
  eps <- 1e-7
  x0 <- seq(min(SCdata[[smoothvar]]), max(SCdata[[smoothvar]]), length.out = n_points)
  x1 <- x0 + eps
  derivative_male <- array(NA, dim = c(n_sites, n_points))
  derivative_female <- array(NA, dim = c(n_sites, n_points))
  FDvalue <- mean(SCdata$meanFD)

  for (j in 1:n_sites){
    siteID.tmp <- unique(gam.data2$siteID)[j]
    command <- "Qua <- getQuantile(mod.tmp, quantile=0.5, term = 'age', fixed.at = list(sex=2, meanFD=FDvalue, randomvar=siteID.tmp), n.points = n_points)"
    eval(parse(text=command))
    derivative_female[j, ] <- (Qua(x1) - Qua(x0)) / eps
    
    command <- "Qua <- getQuantile(mod.tmp, quantile=0.5, term = 'age', fixed.at = list(sex=1, meanFD=FDvalue, randomvar=siteID.tmp), n.points = n_points)"
    eval(parse(text=command))
    derivative_male[j, ] <- (Qua(x1) - Qua(x0)) / eps
  }
  
  derivative <- (derivative_female + derivative_male) /2
  derivative_all <- list(derivative, derivative_female, derivative_male)
  
  return(derivative_all)
}


# Conduct
smoothvar <- "age"
n_points = 1000
n_sites <- nlevels(SCdata.TD.trainset$siteID %>% droplevels())
## TD
# Environment variable
gam.data2 <- SCdata.TD.trainset
gam.data2$randomvar <- as.factor(gam.data2[["siteID"]])
gam.data2 <- as.data.frame(gam.data2)

if (! file.exists(paste0(interfileFolder, "/FirstDeirvative_GAMLSS_Yeo", Yeoresolution,".TDtraining.sum.rds"))){
  DerivativeTD.sum <- mclapply(1:element_num, function(i){
    mod.tmp <- mod.TD.sum[[i]]$mod.tmp
    
    invisible(sumlist <- comuptederivative(smoothvar, mod.tmp, n_points, n_sites))
    sumlist$derivativemean <- colMeans(sumlist[[1]])
    return(sumlist)
  }, mc.cores = 40)
  
  saveRDS(DerivativeTD.sum, paste0(interfileFolder, "/FirstDeirvative_GAMLSS_Yeo", Yeoresolution,".TDtraining.sum.rds"))
}else{
  DerivativeTD.sum <- readRDS(paste0(interfileFolder, "/FirstDeirvative_GAMLSS_Yeo", Yeoresolution,".TDtraining.sum.rds"))
}

## ADHD
# Environment variable
gam.data2 <- SCdata.ADHD
gam.data2$randomvar <- as.factor(gam.data2[["siteID"]])
gam.data2 <- gam.data2 %>% group_by(randomvar) %>%
      filter(n() > 30) %>%
      ungroup()
gam.data2 <- as.data.frame(gam.data2)
n_sites <- nlevels(gam.data2$siteID %>% droplevels())

if (! file.exists(paste0(interfileFolder, "/FirstDeirvative_GAMLSS_Yeo", Yeoresolution,".ADHD.sum.rds"))){
  DerivativeADHD.sum <- mclapply(1:element_num, function(i){
    mod.tmp <- mod.ADHD.sum[[i]]$mod.tmp
    
    invisible(sumlist <- comuptederivative(smoothvar, mod.tmp, n_points, n_sites))
    sumlist$derivativemean <- colMeans(sumlist[[1]])
    return(sumlist)
  }, mc.cores = 40)
  
  saveRDS(DerivativeADHD.sum, paste0(interfileFolder, "/FirstDeirvative_GAMLSS_Yeo", Yeoresolution,".ADHD.sum.rds"))
}else{
  DerivativeADHD.sum <- readRDS(paste0(interfileFolder, "/FirstDeirvative_GAMLSS_Yeo", Yeoresolution,".ADHD.sum.rds"))
}


```


## Step3 Plot the trajectories

```{r}

## 1) Avergae first derivative matrix
DerivativeTD.avg <- lapply(DerivativeTD.sum, function(x) mean(x$derivativemean))
DerivativeTD.avg <- unlist(DerivativeTD.avg)
DerivativeADHD.avg <- lapply(DerivativeADHD.sum, function(x) mean(x$derivativemean))
DerivativeADHD.avg <- unlist(DerivativeADHD.avg)

CentileTD.P50.1 <- lapply(mod.TD.sum, function(x) (x$centiles_strat[[1]][3,1]+x$centiles_strat[[2]][3,1])/2)
CentileTD.P50.1 <- unlist(CentileTD.P50.1)

CentileADHD.P50.1 <- lapply(mod.ADHD.sum, function(x) (x$centiles_strat[[1]][3,1]+x$centiles_strat[[2]][3,1])/2)
CentileADHD.P50.1 <- unlist(CentileADHD.P50.1)

Derivative.df <- data.frame(SClabel= paste0("SC.", 1:element_num), DerivativeTD.avg=DerivativeTD.avg / CentileTD.P50.1, DerivativeADHD.avg=DerivativeADHD.avg / CentileADHD.P50.1)

lmthr <- max(abs(c(Derivative.df$DerivativeTD.avg, Derivative.df$DerivativeADHD.avg)))+1e-5 # 0.03660626
axeslabels=c("VS.P", "SM.A", "VS.C", "SM.B", "DA.A", "DA.B", "FP.C", "DM.C", "VA.A", "TP", "FP.A", "VA.B", "DM.A", "FP.B", "DM.B") 
axeslabelsGap=T
PaletteSet <- list(Name="RdBu", drirection=-1, lmmin = -lmthr, lmmax = lmthr, anglex=0, angley=0,
                   hjustx = 1, hjusty = 1, vjustx = 1, vjusty=0.5)
AvgDerivativeTD <- plotmatrix(dataname="Derivative.df", variable="DerivativeTD.avg", ds.resolution=Yeoresolution.delLM, Pvar=NA, NAcol="white", lmthr=NA, axeslabels=axeslabels, axeslabelsGap=F, linerange_frame=NA, PaletteSet=PaletteSet, Pvar.noFDR=NA)
AvgDerivativeTD

AvgDerivativeADHD <- plotmatrix(dataname="Derivative.df", variable="DerivativeADHD.avg", ds.resolution=Yeoresolution.delLM, Pvar=NA, NAcol="white", lmthr=NA, axeslabels=axeslabels, axeslabelsGap=F, linerange_frame=NA, PaletteSet=PaletteSet, Pvar.noFDR=NA)
AvgDerivativeADHD

ggsave(paste0(FigureFolder, "/Normative_Trajectory/AvgDerivative_TD_Mat.tiff"), AvgDerivativeTD, height = 14, width = 16.5, units = "cm")
ggsave(paste0(FigureFolder, "/Normative_Trajectory/AvgDerivative_ADHD_Mat.tiff"), AvgDerivativeADHD, height = 14, width = 16.5, units = "cm")

## 2) The onset age
Agesequence <- seq(min(SCdata[["age"]]), max(SCdata[["age"]]), length.out = 1000)

# TD
DerivativeTD.df <- do.call(rbind, lapply(DerivativeTD.sum, function(x) x$derivativemean))
DerivativeTD.df <- t(DerivativeTD.df)
DerivativeTD.df <- as.data.frame(DerivativeTD.df)
names(DerivativeTD.df) <- paste0("SC.", 1:element_num)
DerivativeTD.df$age <- Agesequence

DerivativeTD.df.l <- DerivativeTD.df %>% pivot_longer(cols=starts_with("SC."), names_to="label_ID",
                                                      values_to="derivative")

peakchangeTD <- DerivativeTD.df.l %>% group_by(label_ID) %>% 
  summarise(peakchange = mean(age[round(abs(derivative), 4) == max(round(abs(derivative), 4))]),
            onsetage = if (sum(derivative > 0) > 0){min(age[derivative > 0])}else{NA})
peakchangeTD$SCID <- as.numeric(gsub("SC.", "", peakchangeTD$label_ID))
peakchangeTD <- peakchangeTD[order(peakchangeTD$SCID), ]

# ADHD
DerivativeADHD.df <- do.call(rbind, lapply(DerivativeADHD.sum, function(x) x$derivativemean))
DerivativeADHD.df <- t(DerivativeADHD.df)
DerivativeADHD.df <- as.data.frame(DerivativeADHD.df)
names(DerivativeADHD.df) <- paste0("SC.", 1:element_num)
DerivativeADHD.df$age <- Agesequence

DerivativeADHD.df.l <- DerivativeADHD.df %>% pivot_longer(cols=starts_with("SC."), names_to="label_ID",
                                                      values_to="derivative")

peakchangeADHD <- DerivativeADHD.df.l %>% group_by(label_ID) %>% 
  summarise(peakchange = mean(age[round(abs(derivative), 4) == max(round(abs(derivative), 4))]),
            onsetage = if (sum(derivative > 0) > 0){min(age[derivative > 0])}else{NA})
peakchangeADHD$SCID <- as.numeric(gsub("SC.", "", peakchangeADHD$label_ID))
peakchangeADHD <- peakchangeADHD[order(peakchangeADHD$SCID), ]


lmmax <- max(c(peakchangeTD$onsetage, peakchangeADHD$onsetage), na.rm=T)
lmmin <- min(c(peakchangeTD$onsetage, peakchangeADHD$onsetage), na.rm=T)
PaletteSet <- list(Name="BuPu", drirection=1, lmmin = lmmin, lmmax = 15.5, anglex=0, angley=0,
                   hjustx = 1, hjusty = 1, vjustx = 0.5, vjusty=0.5)

Matonset.TD <- plotmatrix(dataname="peakchangeTD", variable="onsetage", ds.resolution=Yeoresolution.delLM, Pvar=NA, NAcol="#4D004B", lmthr=NA, axeslabels=axeslabels, axeslabelsGap=F, linerange_frame=NA, PaletteSet=PaletteSet)
Matonset.TD

Matonset.ADHD <- plotmatrix(dataname="peakchangeADHD", variable="onsetage", ds.resolution=Yeoresolution.delLM, Pvar=NA, NAcol="#4D004B", lmthr=NA, axeslabels=axeslabels, axeslabelsGap=F, linerange_frame=NA, PaletteSet=PaletteSet)
Matonset.ADHD

ggsave(paste0(FigureFolder, "/Normative_Trajectory/Onset_TD_Mat.tiff"), Matonset.TD, height = 14, width = 16, units = "cm")
ggsave(paste0(FigureFolder, "/Normative_Trajectory/Onset_ADHD_Mat.tiff"), Matonset.ADHD, height = 14, width = 16, units = "cm")
# Onset age gap
peakchangeall <- merge(peakchangeTD, peakchangeADHD, by="label_ID")
peakchangeall <- peakchangeall %>% mutate(
  onset.gap = case_when(is.na(onsetage.y) & !is.na(onsetage.x) ~ (15.5-onsetage.x),
                        is.na(onsetage.y) & is.na(onsetage.x) ~ NA,
                        .default = onsetage.y - onsetage.x)
)
# Max >=6.583, min=-2.076
peakchangeall <- peakchangeall[order(peakchangeall$SCID.x), ]
peakchangeall <- peakchangeall %>% mutate(
  NAlabel = case_when(is.na(onset.gap) ~ 0.001,
                      .default = 0.9)
)

PaletteSet <- list(Name="RdBu", drirection=-1, lmmin = -6.59, lmmax = 6.59, anglex=0, angley=0,
                   hjustx = 1, hjusty = 1, vjustx = 0.5, vjusty=0.5)
Matonset.gap <- plotmatrix(dataname="peakchangeall", variable="onset.gap", ds.resolution=Yeoresolution.delLM, Pvar=NA, NAcol="gray", lmthr=NA, axeslabels=axeslabels, axeslabelsGap=F, linerange_frame=NA, PaletteSet=PaletteSet, Pvar.noFDR="NAlabel")
Matonset.gap

ggsave(paste0(FigureFolder, "/Normative_Trajectory/Onset_gap_Mat.tiff"), Matonset.gap, height = 14, width = 16, units = "cm")
SCrankcorr(peakchangeall, "onset.gap", Yeoresolution.delLM)
#   ds.resolution Interest.var r.spearman   p.spearman
# 1            15    onset.gap  0.3859273 2.047686e-05

matindex <- matrix(NA, 15, 15)
matindex[lower.tri(matindex, diag = T)] <- 1:element_num
Type.S <- as.numeric(matindex[1:4, 1:4])
Type.SA <- as.numeric(matindex[5:15, 1:4])
Type.A <- as.numeric(matindex[5:15, 5:15])

Type.S <- Type.S[! is.na(Type.S)]
Type.SA <- Type.SA[! is.na(Type.SA)]
Type.A <- Type.A[! is.na(Type.A)]

peakchangeall$Type <- NA
peakchangeall$Type[Type.S] <- "SS"
peakchangeall$Type[Type.A] <- "AA"
peakchangeall$Type[Type.SA] <- "SA"

peakchangeall.l <- peakchangeall %>% pivot_longer(
  cols = c(onsetage.x, onsetage.y),
  names_to = "group",
  values_to = "onsetage"
)%>%
  mutate(group = ifelse(group == "onsetage.x", "TD", "ADHD"))

# T test for three type
resultdf <- data.frame(Type=c("SS", "AA", "SA"), Tvalue=rep(NA, 3),df=rep(NA, 3), Pvalue=rep(NA, 3), cohend=rep(NA, 3))
n=0
for (t in c("SS", "SA", "SA")){
  df.tmp <- peakchangeall.l %>% filter(Type == t)
  df.tmp$onsetage[is.na(df.tmp$onsetage)] <- 15.5
  ttest.res <- t.test(onsetage ~ group, data=df.tmp, paired = T)
  df <- ttest.res$parameter
  difference <- df.tmp %>% group_by(label_ID) %>% summarise(diff = onsetage[group=="ADHD"]-onsetage[group=="TD"])
  cohend.tmp <- mean(difference$diff) / sd(difference$diff) 
  n=n+1
  resultdf$Tvalue[n] <- ttest.res$statistic
  resultdf$df[n] <- as.numeric(df)
  resultdf$Pvalue[n] <- ttest.res$p.value
  resultdf$cohend[n] <- cohend.tmp
}

print(resultdf)

peakchangeall.l$onsetage[is.na(peakchangeall.l$onsetage)] <- 15.5
stats_df <- peakchangeall.l %>% group_by(Type, group) %>%
  summarise(
    mean_age = mean(onsetage, na.rm = TRUE),
    sd_age = sd(onsetage, na.rm = TRUE),
    .groups = "drop"
  )

stats_df$group <- factor(stats_df$group, levels = c("TD", "ADHD"))
stats_df$Type <- factor(stats_df$Type, levels = c("SS", "SA", "AA"))
barplot <- ggplot(stats_df, aes(x = Type, y = mean_age, fill = group)) +
  geom_col(position = position_dodge(width = 0.8), width = 0.7) +
  geom_errorbar(
    aes(ymin = mean_age - sd_age, ymax = mean_age + sd_age),
    position = position_dodge(width = 0.8),
    width = 0.25,
    color = "black"
  ) +
  coord_cartesian(ylim = c(5, 15)) +
  labs(
    x = "Type of connections",
    y = "Increase onset age (years)",
    fill = NULL
  ) +
  scale_fill_manual(values = c("gray", "#B2182B")) +
  theme_classic()+
  theme(axis.text=element_text(size=18.4, color="black"), 
        axis.title =element_text(size=18.4, color="black"),aspect.ratio = 0.95,
        plot.background=element_rect(fill="white"),
        panel.background=element_rect(fill="white"),
        plot.title = element_text(size=16, hjust = 0.5),
        legend.text = element_text(size=18.5, color="black"), legend.title = element_text(size=20, color="black"))

ggsave(paste0(FigureFolder, "/Normative_Trajectory/Onset_gap_barplot.tiff"), barplot, width=17, height =15.7, units = "cm")


## 3) Developmental trajectories
# TD
CentileP50.TD <- do.call(rbind, lapply(mod.TD.sum, function(x) (x$centiles_strat[[1]][3,]+x$centiles_strat[[2]][3,])/2))

CentileP50.TD <- CentileP50.TD / CentileTD.P50.1

CentileP50.TD <- t(CentileP50.TD)
CentileP50.TD <- as.data.frame(CentileP50.TD)
names(CentileP50.TD) <- paste0("SC.", 1:element_num)
CentileP50.TD$age <- seq(min(SCdata.TD.trainset$age), max(SCdata.TD.trainset$age), length.out=1000)

CentileP50.TD.l <- CentileP50.TD %>% pivot_longer(cols=starts_with("SC."), names_to="label_ID",
                                                      values_to="fit")
CentileP50.TD.l <- CentileP50.TD.l %>% left_join(Derivative.df, join_by("label_ID"=="SClabel"))

# ADHD
CentileP50.ADHD <- do.call(rbind, lapply(mod.ADHD.sum, function(x) (x$centiles_strat[[1]][3,]+x$centiles_strat[[2]][3,])/2))
CentileP50.ADHD <- CentileP50.ADHD / CentileADHD.P50.1
CentileP50.ADHD <- t(CentileP50.ADHD)
CentileP50.ADHD <- as.data.frame(CentileP50.ADHD)
names(CentileP50.ADHD) <- paste0("SC.", 1:element_num)
CentileP50.ADHD$age <- seq(min(gam.data2$age), max(gam.data2$age), length.out=1000)

CentileP50.ADHD.l <- CentileP50.ADHD %>% pivot_longer(cols=starts_with("SC."), names_to="label_ID",
                                                      values_to="fit")
CentileP50.ADHD.l <- CentileP50.ADHD.l %>% left_join(Derivative.df, join_by("label_ID"=="SClabel"))


# TD trajectories
lmthr <- max(abs(c(Derivative.df$DerivativeTD.avg, Derivative.df$DerivativeADHD.avg)))+1e-5
#lmthr <- max(abs(Derivative.df$DerivativeTD.avg))+1e-5
ymin <- min(c(CentileP50.TD.l$fit, CentileP50.ADHD.l$fit))
ymax <- max(c(CentileP50.TD.l$fit, CentileP50.ADHD.l$fit))

TDline <- ggplot(data=CentileP50.TD.l)+
  geom_line(aes(x=age, y=fit, group=label_ID, color=DerivativeTD.avg), linewidth=0.8, alpha=0.8)+
  scale_color_distiller(type="seq", palette = "RdBu",limit=c(-lmthr, lmthr), direction = -1)+
  #scale_y_continuous(limits = c(ymin, ymax))+
  labs(x="Age (years)", y="SC strength (ratio)", color="Average \nderivative")+
  theme_classic()+
  theme(axis.text=element_text(size=20, color="black"), 
        axis.title =element_text(size=20, color="black"),aspect.ratio = 1.05,
        plot.background=element_rect(fill="white"),
        panel.background=element_rect(fill="white"),
        plot.title = element_text(size=16, hjust = 0.5),
        legend.text = element_text(size=18, color="black"), legend.title = element_text(size=20, color="black"))

ggsave(paste0(FigureFolder, "/Normative_Trajectory/P50fitlines_TD.tiff"), TDline, width=19, height =18, units = "cm")


ADHDline <- ggplot(data=CentileP50.ADHD.l)+
  geom_line(aes(x=age, y=fit, group=label_ID, color=DerivativeADHD.avg), linewidth=0.8, alpha=0.8)+
  scale_color_distiller(type="seq", palette = "RdBu",limit=c(-lmthr, lmthr), direction = -1)+
  scale_y_continuous(limits = c(ymin, ymax))+
  labs(x="Age (years)", y="SC strength (ratio)", color="Average \nderivative")+
  theme_classic()+
  theme(axis.text=element_text(size=20, color="black"), 
        axis.title =element_text(size=20, color="black"),aspect.ratio = 1.05,
        plot.background=element_rect(fill="white"),
        panel.background=element_rect(fill="white"),
        plot.title = element_text(size=16, hjust = 0.5),
        legend.text = element_text(size=18, color="black"), legend.title = element_text(size=20, color="black"))

ggsave(paste0(FigureFolder, "/Normative_Trajectory/P50fitlines_ADHD.tiff"), ADHDline, width=19, height =18, units = "cm")

```



## Step4 Plot the deviation


```{r}
deviationSCmat.list <- list()
# ADHD
deviationdf.ADHD <- deviationdf %>% filter(if_TD=="ADHD")

deviationSCmat <- deviationdf.ADHD %>% dplyr::select(ends_with("_deviationZ") & starts_with("SC."))
deviationSCmat.mean <- base::colMeans(deviationSCmat)
deviationSCmat.df <- data.frame(SClabel = names(deviationSCmat), deviationMean = deviationSCmat.mean)
deviationSCmat.list[["ADHD"]] <- deviationSCmat.df

lmthr <- max(abs(deviationSCmat.df$deviationMean))+1e-5 # 0.1312783
axeslabels=c("VS.P", "SM.A", "VS.C", "SM.B", "DA.A", "DA.B", "FP.C", "DM.C", "VA.A", "TP", "FP.A", "VA.B", "DM.A", "FP.B", "DM.B") 
axeslabelsGap=T
PaletteSet <- list(Name="RdBu", drirection=-1, lmmin = -lmthr, lmmax = lmthr, anglex=0, angley=0,hjustx = 1, hjusty = 1, vjustx = 1, vjusty=0.5)

deviationADHD <- plotmatrix(dataname="deviationSCmat.df", variable="deviationMean", ds.resolution=Yeoresolution.delLM, Pvar=NA, NAcol="white", lmthr=NA, axeslabels=axeslabels, axeslabelsGap=F, linerange_frame=NA, PaletteSet=PaletteSet, Pvar.noFDR=NA)
deviationADHD

ggsave(paste0(FigureFolder, "/Normative_Trajectory/MeanDeviation_ADHD_MatAll.tiff"), deviationADHD, height = 14, width = 16, units = "cm")

## Correlation to S-A connectional axis
SCrankcorr(deviationSCmat.df, "deviationMean", Yeoresolution.delLM)
#   ds.resolution  Interest.var r.spearman   p.spearman
# 1            15 deviationMean 0.4364526   6.251687e-07	


# TD
deviationdf.TD <- deviationdf %>% filter(if_TD=="TD")

deviationSCmat <- deviationdf.TD %>% dplyr::select(ends_with("_deviationZ") & starts_with("SC."))
deviationSCmat.mean <- colMeans(deviationSCmat)
deviationSCmat.df <- data.frame(SClabel = names(deviationSCmat), deviationMean = deviationSCmat.mean)
deviationSCmat.list[["TD"]] <- deviationSCmat.df

axeslabels=c("VS.P", "SM.A", "VS.C", "SM.B", "DA.A", "DA.B", "FP.C", "DM.C", "VA.A", "TP", "FP.A", "VA.B", "DM.A", "FP.B", "DM.B") 
axeslabelsGap=T
PaletteSet <- list(Name="RdBu", drirection=-1, lmmin = -lmthr, lmmax = lmthr, anglex=0, angley=0,hjustx = 1, hjusty = 1, vjustx = 1, vjusty=0.5)

deviationTD <- plotmatrix(dataname="deviationSCmat.df", variable="deviationMean", ds.resolution=Yeoresolution.delLM, Pvar=NA, NAcol="white", lmthr=NA, axeslabels=axeslabels, axeslabelsGap=F, linerange_frame=NA, PaletteSet=PaletteSet, Pvar.noFDR=NA)
deviationTD

ggsave(paste0(FigureFolder, "/Normative_Trajectory/MeanDeviation_TD_MatAll.tiff"), deviationTD, height = 14, width = 16, units = "cm")

## Correlation to S-A connectional axis
SCrankcorr(deviationSCmat.df, "deviationMean", Yeoresolution.delLM)
#   ds.resolution  Interest.var r.spearman   p.spearman
# 1            15 deviationMean  0.125237  0.1729092	


### Interaction effect
deviationSCmat.list[[1]]$if_TD <- "ADHD"
deviationSCmat.list[[2]]$if_TD <- "TD"

deviationSCmat.all <- do.call(rbind, deviationSCmat.list)
SCrankdf <- SCrankcorr(deviationSCmat.df, "deviationMean", Yeoresolution.delLM, T)
SCrankdf$SClabel <- paste0("SC.", 1:element_num, "_deviationZ")
SCrankdf$deviationMean <- NULL
deviationSCmat.all <- deviationSCmat.all %>% left_join(SCrankdf, by="SClabel")
deviationSCmat.all$if_TD <- factor(deviationSCmat.all$if_TD, levels=c("TD", "ADHD"))
intmodel <- lm(deviationMean ~ SCrank*if_TD, data = deviationSCmat.all)
nullmodel <- lm(deviationMean ~ SCrank+if_TD, data = deviationSCmat.all)
summary(intmodel)
# Coefficients:
#                  Estimate Std. Error t value Pr(>|t|)    
# (Intercept)    -0.0228464  0.0074264  -3.076 0.002343 ** 
# SCrank          0.0006229  0.0001065   5.848 1.65e-08 ***
# if_TDTD         0.0122839  0.0105026   1.170 0.243338    
# SCrank:if_TDTD -0.0005065  0.0001507  -3.362 0.000903 ***
# ---
# Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
# 
# Residual standard error: 0.04042 on 236 degrees of freedom
# Multiple R-squared:  0.1683,	Adjusted R-squared:  0.1578 
# F-statistic: 15.92 on 3 and 236 DF,  p-value: 1.844e-09
anova(nullmodel, intmodel) # F = 11.301, P=0.0009034

scatter <- ggplot(data=deviationSCmat.all)+
    geom_point(aes(x=SCrank, y=deviationMean, color=if_TD), size=1.8)+
    geom_smooth(aes(x=SCrank, y=deviationMean, fill=if_TD), method ="lm", color="black", alpha=0.3, linewidth=1.2)+
    scale_color_manual(values = c("gray", "#B2182B"))+
    scale_fill_manual(values = c("gray", "#B2182B"))+
    labs(x="S-A connectional axis rank", y="Average SC deviation", color=NULL, fill=NULL)+
    #scale_y_continuous(breaks = c(-0.015, -0.010, -0.005,0, 0.005, 0.010), labels = c(-15,-10, -5, 0, 5, 10))+
    scale_y_continuous(breaks = c(-0.1, 0, 0.1))+
    scale_x_continuous(breaks = c(0,20,40,60,80,100,120))+
    theme_classic()+
    theme(axis.text=element_text(size=20.3, color="black"), 
        axis.title =element_text(size=20.3, color="black"),aspect.ratio = 1.07,
        plot.background=element_rect(fill="white"),
        panel.background=element_rect(fill="white"),
        plot.title = element_text(size=20, hjust = 0.5),
        legend.text = element_text(size=20, color="black"), legend.title = element_text(size=20, color="black"))
ggsave(paste0(FigureFolder, '/Normative_Trajectory/AvgDeivation_SCrankcorr_Interaction.tiff'),scatter, width=18.1, height =16.5, units = "cm")

## Select a random edge

for (i in c(3,64,65,84,96,98,109,110,117,119)){
  SClabel = paste0("SC.", i)
  centile.tmp <- mod.TD.sum[[i]]$centiles_strat
  centile.tmp <- (centile.tmp[[1]] + centile.tmp[[2]]) / 2
  centile.tmp <- t(centile.tmp)
  centile.tmp <- as.data.frame(centile.tmp)
  names(centile.tmp) <- c("P2.5", "P50", "P97.5")
  
  centile.tmp$age <- CentileP50.TD$age
  centile.tmp <- centile.tmp %>% pivot_longer(starts_with("P"), names_to="Percentile", values_to = "fit")
  centile.tmp$Percentile <- factor(centile.tmp$Percentile, levels=c("P97.5", "P50", "P2.5"))
  
  pointdf.ADHD <- deviationdf.ADHD[,c("age", paste0(SClabel, "_standard"), "subID")]
  pointdf.TDtest <- deviationdf[deviationdf$if_TD=="TD",c("age", paste0(SClabel, "_standard"))]
  
  TDtest.standard <- pointdf.TDtest[[paste0(SClabel, "_standard")]]
  ADHD.standard <- pointdf.ADHD[[paste0(SClabel, "_standard")]]
  
  Figtmp <- ggplot()+
    geom_hex(data=pointdf.TDtest, aes(x=age, y=TDtest.standard), bins=20)+
    scale_fill_gradientn(colors = c("#F0F0F0", "#D9D9D9", "#BDBDBD", "#969696")) +
    geom_line(data=pointdf.ADHD, aes(x=age, y=ADHD.standard, group=subID), color="#FDDBC7")+
    geom_point(data=pointdf.ADHD, aes(x=age, y=ADHD.standard), color="#B2182B")+
    geom_line(data=centile.tmp, aes(x=age, y=fit, group=Percentile, linetype=Percentile), color="black", linewidth=1)+
    scale_linetype_manual(values=c("dashed", "solid", "dashed"))+
    scale_y_continuous(breaks = c(0,1,2,3,4))+
    labs(x="Age (years)", y="SC strength")+
    theme_classic()+
    theme(axis.text=element_text(size=20, color="black"), 
          axis.title =element_text(size=20, color="black"),aspect.ratio = 1.13,
          plot.background=element_rect(fill="white"),
          panel.background=element_rect(fill="white"),
          plot.title = element_text(size=20, hjust = 0.5),
          legend.text = element_text(size=20, color="black"), legend.title = element_text(size=20, color="black"))
  
  ggsave(paste0(FigureFolder, "/Normative_Trajectory/NormCurve_", SClabel,"_ADHDDeviation.tiff"), Figtmp, width=17, height =16, units = "cm")
  
}

### WB mean SC
SClabel = "WB_SCmean"
centile.tmp <- mod.training.WB$centiles_strat
centile.tmp <- (centile.tmp[[1]] + centile.tmp[[2]]) / 2
centile.tmp <- t(centile.tmp)
centile.tmp <- as.data.frame(centile.tmp)
names(centile.tmp) <- c("P2.5", "P50", "P97.5")

centile.tmp$age <- CentileP50.TD$age
centile.tmp <- centile.tmp %>% pivot_longer(starts_with("P"), names_to="Percentile", values_to = "fit")
centile.tmp$Percentile <- factor(centile.tmp$Percentile, levels=c("P97.5", "P50", "P2.5"))

pointdf.ADHD <- deviationdf.ADHD[,c("age", paste0(SClabel, "_standard"), "subID")]
pointdf.TDtest <- deviationdf[deviationdf$if_TD=="TD",c("age", paste0(SClabel, "_standard"))]

TDtest.standard <- pointdf.TDtest[[paste0(SClabel, "_standard")]]
ADHD.standard <- pointdf.ADHD[[paste0(SClabel, "_standard")]]

Figtmp <- ggplot()+
  #geom_hex(data=pointdf.TDtest, aes(x=age, y=TDtest.standard), bins=15)+
  geom_density_2d_filled(data=pointdf.TDtest, aes(x=age, y=TDtest.standard),contour_var = "count",bins=50, alpha=1, show.legend=F)+
  scale_fill_manual(values=COL1(sequential="Greys", n=55)[c(1,6:55)])+
  #scale_fill_gradient(low = "grey90",  high = "grey20", na.value = "white") +
  #scale_color_distiller(type="seq", palette = "RdGy", direction = -1)+
  geom_line(data=pointdf.ADHD, aes(x=age, y=ADHD.standard, group=subID), color="#FDDBC7")+
  geom_point(data=pointdf.ADHD, aes(x=age, y=ADHD.standard), color="#B2182B", size=1.8)+
  geom_line(data=centile.tmp, aes(x=age, y=fit, group=Percentile, linetype=Percentile), color="black", linewidth=1)+
  scale_linetype_manual(values=c("dashed", "solid", "dashed"))+
  #scale_y_continuous(breaks = c(0,1,2,3,4))+
  labs(x="Age (years)", y="SC strength")+
  theme_classic()+
  theme(axis.text=element_text(size=20.3, color="black"), 
        axis.title =element_text(size=20.3, color="black"),aspect.ratio = 0.99,
        plot.background=element_rect(fill="white"),
        panel.background=element_rect(fill="white"),
        plot.title = element_text(size=20, hjust = 0.5),
        legend.text = element_text(size=20, color="black"), legend.title = element_text(size=20, color="black"))

ggsave(paste0(FigureFolder, "/Normative_Trajectory/NormCurve_", SClabel,"_ADHDDeviation.tiff"), Figtmp, width=19, height =17, units = "cm")



## Draw a schema

pointdf.TDtest2 <- deviationdf %>% filter(if_TD=="TD", abs(WB_SCmean_deviationZ) <2.3) 
pointdf.TDtest2 <- pointdf.TDtest2[sample(1:100, 100, F),]

pointdf.ADHDtest2 <- deviationdf %>% filter(if_TD=="ADHD") 
pointdf.ADHDtest3 <- deviationdf %>% filter(if_TD=="ADHD", WB_SCmean_deviationZ > 1.96) 
pointdf.ADHDtest2 <- pointdf.ADHDtest2[sample(1:90, 90, F),]
pointdf.ADHDtest2 <- rbind(pointdf.ADHDtest2, pointdf.ADHDtest3[sample(1:10, 10, F),])

Figtmp <- ggplot()+
  geom_line(data=pointdf.TDtest2, aes(x=age, y=WB_SCmean_standard, group=subID), color="#D9D9D9")+
  geom_point(data=pointdf.TDtest2, aes(x=age, y=WB_SCmean_standard), color="#969696", size=1.8)+
  geom_line(data=pointdf.ADHDtest2, aes(x=age, y=WB_SCmean_standard, group=subID), color="#FDDBC7")+
  geom_point(data=pointdf.ADHDtest2, aes(x=age, y=WB_SCmean_standard), color="#B2182B", size=1.8)+
  geom_line(data=centile.tmp, aes(x=age, y=fit, group=Percentile, linetype=Percentile), color="black", linewidth=0.8)+
  scale_linetype_manual(values=c("dashed", "solid", "dashed"))+
  #scale_y_continuous(breaks = c(0,1,2,3,4))+
  labs(x="Age (years)", y="SC strength")+
  theme_classic()+
  theme(axis.text=element_text(size=20.3, color="black"), 
        axis.title =element_text(size=20.3, color="black"),aspect.ratio = 0.88,
        plot.background=element_rect(fill="white"),
        panel.background=element_rect(fill="white"),
        plot.title = element_text(size=20, hjust = 0.5),
        legend.text = element_text(size=20, color="black"), legend.title = element_text(size=20, color="black"))

ggsave(paste0(FigureFolder, "/Normative_Trajectory/NormCurve_schema.tiff"), Figtmp, width=14, height =13, units = "cm")
ggsave(paste0(FigureFolder, "/Normative_Trajectory/NormCurve_schema.svg"), Figtmp, width=14, height =13, units = "cm")

### SC deviation examples
## Randomly select 3 subjects
# ADHD
mat1 <- deviationdf.ADHD %>% filter(round(age) == 9) %>% dplyr::select(ends_with("_deviationZ") & starts_with("SC."))
mat1 <- mat1[28,]
mat2 <- deviationdf.ADHD %>% filter(round(age) == 9) %>% dplyr::select(ends_with("_deviationZ") & starts_with("SC."))
mat2 <- mat2[56,]
mat3 <- deviationdf.ADHD %>% filter(round(age) == 10) %>% dplyr::select(ends_with("_deviationZ") & starts_with("SC."))
mat3 <- mat3[28,]

df <- data.frame(mat1=as.numeric(mat1), mat2=as.numeric(mat2),mat3=as.numeric(mat3))

maxvalue <- max(c(df$mat1, df$mat2, df$mat3))

PaletteSet <- list(Name="RdBu", drirection=-1, lmmin = -maxvalue, lmmax = maxvalue, anglex=0, angley=0,
                   hjustx = 1, hjusty = 1, vjustx = 1, vjusty=0.5)
axeslabels=c("VS.P", "SM.A", "VS.C", "SM.B", "DA.A", "DA.B", "FP.C", "DM.C", "VA.A", "TP", "FP.A", "VA.B", "DM.A", "FP.B", "DM.B") 

plotmat1 <- plotmatrix("df", variable="mat1", ds.resolution=15, Pvar=NA, NAcol="white", lmthr=NA, 
                       axeslabels=axeslabels, axeslabelsGap=F, linerange_frame=NA, PaletteSet=PaletteSet, Pvar.noFDR=NA)

plotmat2 <- plotmatrix("df", variable="mat2", ds.resolution=15, Pvar=NA, NAcol="white", lmthr=NA, 
                       axeslabels=axeslabels, axeslabelsGap=F, linerange_frame=NA, PaletteSet=PaletteSet, Pvar.noFDR=NA)

plotmat3 <- plotmatrix("df", variable="mat3", ds.resolution=15, Pvar=NA, NAcol="white", lmthr=NA, 
                       axeslabels=NULL, axeslabelsGap=F, linerange_frame=NA, PaletteSet=PaletteSet, Pvar.noFDR=NA)


ggsave(paste0(FigureFolder, "/Normative_Trajectory/SCdeviationADHD1.tiff"), plotmat1, height = 12.5, width = 14, units = "cm")
ggsave(paste0(FigureFolder, "/Normative_Trajectory/SCdeviationADHD2.tiff"), plotmat2, height = 12.5, width = 14, units = "cm")
ggsave(paste0(FigureFolder, "/Normative_Trajectory/SCdeviationADHD3.tiff"), plotmat3, height = 12.5, width = 14, units = "cm")

# TD
mat1 <- deviationdf.TD %>% filter(round(age) == 9, abs(WB_SCmean_deviationZ) <1.96) %>% dplyr::select(ends_with("_deviationZ") & starts_with("SC."))
mat1 <- mat1[28,]
mat2 <- deviationdf.TD %>% filter(round(age) == 9, abs(WB_SCmean_deviationZ) <1.96) %>% dplyr::select(ends_with("_deviationZ") & starts_with("SC."))
mat2 <- mat2[56,]
mat3 <- deviationdf.TD %>% filter(round(age) == 10, abs(WB_SCmean_deviationZ) <1.96) %>% dplyr::select(ends_with("_deviationZ") & starts_with("SC."))
mat3 <- mat3[28,]

df <- data.frame(mat1=as.numeric(mat1), mat2=as.numeric(mat2),mat3=as.numeric(mat3))

PaletteSet <- list(Name="RdBu", drirection=-1, lmmin = -maxvalue, lmmax = maxvalue, anglex=0, angley=0,
                   hjustx = 1, hjusty = 1, vjustx = 1, vjusty=0.5)
axeslabels=c("VS.P", "SM.A", "VS.C", "SM.B", "DA.A", "DA.B", "FP.C", "DM.C", "VA.A", "TP", "FP.A", "VA.B", "DM.A", "FP.B", "DM.B") 

plotmat1 <- plotmatrix("df", variable="mat1", ds.resolution=15, Pvar=NA, NAcol="white", lmthr=NA, 
                       axeslabels=axeslabels, axeslabelsGap=F, linerange_frame=NA, PaletteSet=PaletteSet, Pvar.noFDR=NA)

plotmat2 <- plotmatrix("df", variable="mat2", ds.resolution=15, Pvar=NA, NAcol="white", lmthr=NA, 
                       axeslabels=NULL, axeslabelsGap=F, linerange_frame=NA, PaletteSet=PaletteSet, Pvar.noFDR=NA)

plotmat3 <- plotmatrix("df", variable="mat3", ds.resolution=15, Pvar=NA, NAcol="white", lmthr=NA, 
                       axeslabels=NULL, axeslabelsGap=F, linerange_frame=NA, PaletteSet=PaletteSet, Pvar.noFDR=NA)


ggsave(paste0(FigureFolder, "/Normative_Trajectory/SCdeviationTD1.tiff"), plotmat1, height = 12.5, width = 14, units = "cm")
ggsave(paste0(FigureFolder, "/Normative_Trajectory/SCdeviationTD2.tiff"), plotmat2, height = 12.5, width = 14, units = "cm")
ggsave(paste0(FigureFolder, "/Normative_Trajectory/SCdeviationTD3.tiff"), plotmat3, height = 12.5, width = 14, units = "cm")



```
